# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is designed to improve the management of project and experience information for your portfolio website, ensuring a seamless rendering and effortless updating process. Additionally, it facilitates interaction with Firebase Storage, allowing you to efficiently load and display icons and images, enhancing the overall visual experience of your portfolio.",
        "title": "Portfolio server",
        "contact": {
            "name": "Yago Faran",
            "email": "yagofaran@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/images/icons/{name}": {
            "get": {
                "description": "This endpoint returns the icon file by its name",
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/svg+xml",
                    "image/gif"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Get icon file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
...
}